api-agents-whatsapp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "api-agents-whatsapp"
version = "0.1.0"
description = "API de Agentes para gerar mensagens do WhatsApp"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "agno>=0.5.0",
    "openai>=1.11.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.1",
    "mcp",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.12.0",
    "ruff>=0.1.11",
]

[project.scripts]
api-agents = "api_agents_whatsapp.main:main"